feat: tooltip provider
