+++ packages/raystack/components/tooltip/index.tsx
@@ -1,3 +1,1 @@
-export { Tooltip, TooltipProvider } from "./tooltip"
-
-// Todo: fix the light border on top of Arrow. 
+export { Tooltip } from './tooltip';
